+++ packages/webcomponents/src/components/business-forms/business-form-stepped/additional-questions/business-additional-questions-form-step.tsx
@@ -1,10 +1,9 @@
-import { Component, Host, Method, Prop, State, h, Event, EventEmitter } from '@stencil/core';
-import { FormController } from '../../form/form';
-import Api, { IApiResponse } from '../../../api/Api';
-import { Address, BusinessFormServerErrors, IAddress, IBusiness } from '../../../api/Business';
-import { parseAddressInfo } from '../helpers';
-import legalAddressSchema from '../../business-form/legal-address-form/legal-address-form-schema';
-import { config } from '../../../../config';
+import { Component, Host, h, Prop, State, Method, Event, EventEmitter } from '@stencil/core';
+import { FormController } from '../../../form/form';
+import { BusinessFormServerErrors, IBusiness } from '../../../../api/Business';
+import { Api, IApiResponse } from '../../../../api';
+import { config } from '../../../../../config';
+import { additionQuestionsSchema } from '../../business-form/business-form-schema';
 
 /**
  * @exportedPart label: Label for inputs
@@ -12,19 +11,18 @@ import { config } from '../../../../config';
  * @exportedPart input-invalid: Invalid state for inputfs
  */
 @Component({
-  tag: 'justifi-legal-address-form-step',
-  styleUrl: 'legal-address-form-step.scss',
+  tag: 'justifi-additional-questions-form-step',
 })
-export class LegalAddressFormStep {
+export class AdditionalQuestionsFormStep {
   @Prop() authToken: string;
   @Prop() businessId: string;
   @State() formController: FormController;
   @State() errors: any = {};
-  @State() legal_address: IAddress = {};
+  @State() additional_questions: any = {};
   @Event({ bubbles: true }) submitted: EventEmitter<{ data?: any }>;
   @Event() formLoading: EventEmitter<boolean>;
   @Event() serverError: EventEmitter<{ data: any, message: BusinessFormServerErrors }>;
-
+  
   private api: any;
 
   get businessEndpoint() {
@@ -35,8 +33,8 @@ export class LegalAddressFormStep {
     this.formLoading.emit(true);
     try {
       const response: IApiResponse<IBusiness> = await this.api.get(this.businessEndpoint);
-      this.legal_address = new Address(response.data.legal_address);
-      this.formController.setInitialValues({ ...this.legal_address });
+      this.additional_questions = response.data.additional_questions;
+      this.formController.setInitialValues(this.additional_questions);
     } catch (error) {
       this.serverError.emit({ data: error, message: BusinessFormServerErrors.fetchData });
     } finally {
@@ -47,8 +45,8 @@ export class LegalAddressFormStep {
   private sendData = async (onSuccess?: () => void) => {
     this.formLoading.emit(true);
     try {
-      const payload = parseAddressInfo(this.formController.values.getValue());
-      const response = await this.api.patch(this.businessEndpoint, JSON.stringify({ legal_address: payload}));
+      const payload = this.formController.values.getValue();
+      const response = await this.api.patch(this.businessEndpoint, JSON.stringify({ additional_questions: payload}));
       this.handleResponse(response, onSuccess);
     } catch (error) {
       this.serverError.emit({ data: error, message: BusinessFormServerErrors.patchData });
@@ -77,14 +75,14 @@ export class LegalAddressFormStep {
     if (!this.authToken) console.error(missingAuthTokenMessage);
     if (!this.businessId) console.error(missingBusinessIdMessage);
 
-    this.formController = new FormController(legalAddressSchema);
+    this.formController = new FormController(additionQuestionsSchema);
     this.api = Api(this.authToken, config.proxyApiOrigin);
     this.fetchData();
   }
 
   componentDidLoad() {
     this.formController.values.subscribe(values =>
-      this.legal_address = { ...values }
+      this.additional_questions = { ...values }
     );
     this.formController.errors.subscribe(
       errors => (this.errors = { ...errors }),
@@ -99,70 +97,58 @@ export class LegalAddressFormStep {
   }
 
   render() {
-    const legalAddressDefaultValue =
+    const additionalQuestionsDefaultValue =
       this.formController.getInitialValues();
 
     return (
       <Host exportparts="label,input,input-invalid">
         <form>
           <fieldset>
-            <legend>Business Legal Address</legend>
+            <legend>Additional Questions</legend>
             <hr />
-            <div class="row gx-2 gy-2">
+            <div class="row gy-3">
               <div class="col-12">
-                <form-control-text
-                  name="line1"
-                  label="Legal Address"
+                <form-control-monetary
+                  name="business_revenue"
+                  label="Business Revenue"
                   inputHandler={this.inputHandler}
-                  defaultValue={legalAddressDefaultValue?.line1}
-                  error={this.errors?.line1}
+                  error={this.errors?.business_revenue}
+                  defaultValue={additionalQuestionsDefaultValue?.business_revenue}
                 />
               </div>
               <div class="col-12">
-                <form-control-text
-                  name="line2"
-                  label="Address Line 2"
-                  inputHandler={(name, value) => this.inputHandler(name, value)}
-                  defaultValue={legalAddressDefaultValue?.line2}
+                <form-control-monetary
+                  name="business_payment_volume"
+                  label="Business Payment Volume"
+                  inputHandler={this.inputHandler}
+                  error={this.errors?.business_payment_volume}
+                  defaultValue={
+                    additionalQuestionsDefaultValue?.business_payment_volume
+                  }
                 />
               </div>
               <div class="col-12">
-                <form-control-text
-                  name="city"
-                  label="City"
+                <form-control-monetary
+                  name="business_dispute_volume"
+                  label="Business Dispute Volume"
                   inputHandler={this.inputHandler}
-                  defaultValue={legalAddressDefaultValue?.city}
-                  error={this.errors?.city}
+                  error={this.errors?.business_dispute_volume}
+                  defaultValue={
+                    additionalQuestionsDefaultValue?.business_dispute_volume
+                  }
                 />
               </div>
               <div class="col-12">
-                <form-control-text
-                  name="state"
-                  label="State"
+                <form-control-monetary
+                  name="business_receivable_volume"
+                  label="Business Receivable Volume"
                   inputHandler={this.inputHandler}
-                  defaultValue={legalAddressDefaultValue?.state}
-                  error={this.errors?.state}
-                />
-              </div>
-              <div class="col-12">
-                <form-control-number
-                  name="postal_code"
-                  label="Postal Code"
-                  inputHandler={this.inputHandler}
-                  defaultValue={legalAddressDefaultValue?.postal_code}
-                  error={this.errors?.postal_code}
-                />
-              </div>
-              <div class="col-12">
-                <form-control-select
-                  name="country"
-                  label="Country"
-                  options={[{ label: 'United States', value: 'USA' }]}
-                  inputHandler={this.inputHandler}
-                  defaultValue={legalAddressDefaultValue?.country}
-                  error={this.errors?.country}
-                  // just for now so we skip handling country specificities
-                  disabled={true}
+                  error={
+                    this.errors?.business_receivable_volume
+                  }
+                  defaultValue={
+                    additionalQuestionsDefaultValue?.business_receivable_volume
+                  }
                 />
               </div>
             </div>
