+++ packages/webcomponents/src/components/business-forms/business-form-stepped/business-core-info/business-core-info-form-step.scss
