+++ packages/webcomponents/src/components/business-forms/business-form-stepped/business-core-info/business-core-info-form-step.tsx
@@ -1,13 +1,13 @@
 import { Component, Host, h, Prop, State, Method, Event, EventEmitter } from '@stencil/core';
 import { BusinessStructureOptions, BusinessTypeOptions } from '../business-form-schema';
-import { FormController } from '../../form/form';
-import { PHONE_MASKS, TAX_ID_MASKS } from '../../../utils/form-input-masks';
-import { BusinessFormServerErrors, CoreBusinessInfo, IBusiness, ICoreBusinessInfo } from '../../../api/Business';
-import { Api, IApiResponse } from '../../../api';
+import { FormController } from '../../../form/form';
+import { PHONE_MASKS, TAX_ID_MASKS } from '../../../../utils/form-input-masks';
+import { BusinessFormServerErrors, CoreBusinessInfo, IBusiness, ICoreBusinessInfo } from '../../../../api/Business';
+import { Api, IApiResponse } from '../../../../api';
 import { coreInfoSchema } from '../../business-form/business-form-schema';
-import { config } from '../../../../config';
+import { config } from '../../../../../config';
 import { parseCoreInfo } from '../helpers';
-import { flattenNestedObject } from '../../../utils/utils';
+import { flattenNestedObject } from '../../../../utils/utils';
 
 /**
  *
