+++ packages/webcomponents/src/components/business-forms/business-form-stepped/business-form-stepped.scss
