+++ packages/webcomponents/src/components/business-forms/business-form-stepped/business-representative/business-representative-form-step.scss
