+++ packages/webcomponents/src/components/business-forms/business-form-stepped/helpers.ts
