+++ packages/webcomponents/src/components/business-forms/business-form/business-generic-info/business-generic-info.scss
